feat: Completed Calculator exercise

+++ Assignments/Day_10/calculator/main.py
@@ -30,7 +30,7 @@ operations = {
 
 def calculator():
     # Prompt the user to enter the first number
-    num_1 = int(input("What's the first number? "))
+    num_1 = float(input("What's the first number? "))
 
     # Display the available operators
     for operator, operation_function in operations.items():
@@ -45,7 +45,7 @@ def calculator():
         operator = input("Pick an operation: ")
 
         # Prompt the user to enter the next number
-        num_2 = int(input("What's the next number? "))
+        num_2 = float(input("What's the next number? "))
 
         # Retrieve the corresponding arithmetic operation function based on the selected operator
         operation_function = operations[operator]
